[Case Note Search] Improve error handling. (Recidiviz/recidiviz-data#33021)
## Description of the change

Improve error handling.

@lilidworkin pointed out some issues with how we are handling errors.
This PR addresses them:

>If case_note_search errors, then case_note_response will have no
results and an error field (which the FE renders as unexpected error),
but we don't set the status_code of the response to 500. I think this
explains why I wasn't seeing some errors appear in the GCP logs. Can we
look at case_note_response before returning it and set the status_code
to 500 if error is populated?

>I noticed that if record_case_note_search fails (which is the method
that logs the search to the DB), we do set the status_code to 500, and
we return no results, but I think this shouldn't be the behavior. If we
have results, we should return them, even if for some reason we weren't
able to log them. So in this case, let's actually not set the
status_code to 500 -- let's return the results we have, but log to
Sentry so that we know there was an error.

## Testing
Tested this locally to confirm that 1) we return a 500 now if case note
search errors (this should trigger GCP error logging now) and 2) we
still return the case note responses even if we have a bug when logging
the search request/results to our BQ table.

---------

Co-authored-by: Roshan A <roshan@recidiviz.org>
GitOrigin-RevId: f0381121c7697f163eff2ac1e6a07fc645d48317