[Insights] Add earliest_person_assignment_date to supervision_officers (Recidiviz/recidiviz-data#33012)
## Description of the change

For Action Strategies we need to know if an officer first had a caseload
within the last 15 months. I thought about trying to use
`SupervisionClientEvents` for this, but it doesn't seem to completely
fulfill the requirement. For example, if an officer had a client WITHOUT
any events prior to 15 months but their first client with an event was
within 15 months.

So I decided to add `earliest_person_assignment_date` to
`supervision_officers` which may be a naive approach? I will follow this
with another PR to add the column to the postgres schema and model once
the column is in the prod view.

I ran this in the jovergaag_officers sandbox, and also ran
`compare_views` on `supervision_officer_supervisors` to make sure that
it didn't change anything there since supervisors and officers share a
staff query template.

I also threw in a small documentation typo fix in
`load_views_to_sandbox`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 257fe2dc51565c6889aaf2e042522d2992f72808