[Normalization] Fix bug in EntityBackedgeValidator check (Recidiviz/recidiviz-data#33050) (Recidiviz/recidiviz-data#33058)
## Description of the change

In the legacy ingest schema a `StateCharge` can have an empty
`incarceration_sentences` OR `supervision_sentences` backedge list if
it's only connected to one (the common case). This fixes the validation
check to check for truly null backedges but allow empty list backedges.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |
## Related issues

Part of Recidiviz/recidiviz-data#29517

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 07dfd32e6c51cf5b647d37eddb90240b56fd4d1c