[US_AZ] Refactor incarceration periods (Recidiviz/recidiviz-data#32776)
## Description of the change

This PR changes the incarceration periods view in AZ to be simpler, more
similar in structure to the supervision periods view for continuity, and
more accurate.

Before this refactor, the lowest error rate for the
`incarceration_population_by_custody_level_external_comparison`
validation in AZ was 38%. After this change, the lowest error rate is
0.8%, with the lowest failing rate being 3%.

The highest error rate is still 1.0 because we do not yet categorize any
portion of the population as having a "custody level" of Reception or
Inpatient Care Unit in the internal data, but those categorizations do
exist in the external data.

Non-100% error rates dropped from an average of 78.9% to an average of
12% after this change.

Tests:
- [x] ingest view query is deterministic
- [x] external IDs generated by ingest view & mapping are unique
- [x] run_mappings script passes
- [x] sandbox ingest
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-08-30_12_28_35-12407127311768006076;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
succeeds
- [x] normalization
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-08-30_13_01_22-12947030758485036002;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
succeeds
- [x] metrics
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-08-30_13_26_47-10377729299912659504;graphView=1?project=recidiviz-staging)
succeeds
- [x] custody_level_external_comparison validation improves

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4a9370503faf0e1d671de79a1236244f8550bea2