[Case Note Search] Sanitize exact search query terms. (Recidiviz/recidiviz-data#33033)
## Description of the change

Sanitize exact search query terms.

During Wednesday's bug bash, any query terms with apostrophes broke the
BigQuery exact search request. This is a result of how the BQ query is
constructed (the query term is wrapped in apostrophes).

```
    query = f"""
    select *
    from {CASE_NOTES_BQ_TABLE_NAME}
    where LOWER(JSON_EXTRACT_SCALAR(JsonData, '$.note_body')) LIKE LOWER('%{query_term}%')
    """
```

The best practice is to parametrize the queries. However
pandas_gbq.read_gbq() does not support parameterized queries, so we are
moving to simply using a bigquery client.

## Testing
Tested this locally and now apostrophe matching is working as expected,
giving us exact string matches for terms like "her's".

GitOrigin-RevId: 4861fdc634d4a09f071e9bb5996ffb6a38ffb4dd