[US_AZ] Add full name as staff external ID type (Recidiviz/recidiviz-data#32902)
## Description of the change

We learned that the current external ID we use for staff in AZ is not
stable over time, and there is no table connecting a staff member to all
of the IDs they have had over the course of their tenure at ADCRR. Since
staff can be connected to past caseloads using IDs that are no longer
active, to be able to construct an accurate historical image of
supervision officer assignments, we have to be able to attach all of the
relevant `US_AZ_PERSON_ID` external IDs to the same staff member.

Since staff members regularly share generic email addresses in AZ, the
most reliable approach available to do this is using the staff member's
full name. This PR creates a new external ID type `US_AZ_STAFF_NAME`
that concatenates the most recent first, middle, and last name for each
staff member from the `PERSON` raw data table. It then iterates through
all of the `US_AZ_PERSON_ID` values associated with that full name, and
attaches them to the same `staff_id` on the `state_staff` entity.

The pros of this approach are that it lets us connect, with at least
some degree of confidence, a staff member to their historical caseload.
Without this, each ID a staff member has ever had would be tracked as a
distinct person.

The cons of this approach are that we are likely conflating staff
members over time who had the same name. Compiling IDs by full name
assumes that there has been exactly one person with each name who worked
for the department; in cases where that is not true, we will be creating
one US_AZ_STAFF_NAME ID attached to every ID anyone with that name has
had historically.

Tests:
- [x] ingest view query is deterministic
- [x] sandbox ingest
[pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-08-30_13_33_28-4715009413632187550;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
succeeds

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#32901

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 284eae7069f5bf2fe0bd7447e74ce94545d9a15e