Add query to do full diff of raw data tables (Recidiviz/recidiviz-data#33031)
## Description of the change

    Add query to compare raw data between tables.
Finds comparison table rows that have no exact match in the source table
and
source table rows that have no exact match in the comparison table, then
attempts to match rows by joining based on primary keys and
update_datetime.
If the file tag has no documented primary key, all columns are used as
the primary key.
If a row in the source table has no exact primary key match in the
comparison table,
it is returned with all of the comparison columns as None, and vice
versa.
    The first 500 results are returned ordered by update_datetime.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#32737

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f2200f865e20a92d8803ab4d35e79171fa626c1d