[US_IX] Create `state_charge` entity for state sentence refactor  (Recidiviz/recidiviz-data#32147)
## Description of the change

This PR hydrates a new state_charge entity for US_IX

Go here to see the sandbox state_charge_v2 entity:
[`SELECT * FROM
recidiviz-staging.santy_charge_us_ix_state.state_charge_v2`
](https://console.cloud.google.com/bigquery?ws=!1m5!1m4!4m3!1srecidiviz-staging!2ssanty_charge_us_ix_state!3sstate_charge_v2)

New state garge entity  has 344977 charges
New state garge entity  has 349524 charges
Difference of 4547 charges, or ~1.3% of total charges. I believe this
difference is that we are now checking that a charge is not assossiated
to an error correction sentence since that is the only difference in the
queries of the old and new state_charge entity.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#32066

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f8f9b7a6d2960e355b95610c921e9b013d7f5915