[Normalization] Properly merge all multi-parent legacy normalized charges (Recidiviz/recidiviz-data#33070)
## Description of the change

Partially reverts Recidiviz/recidiviz-data#32797 because it turns out that the constraint I was
trying to enforce (no multi-parent legacy charges) is not actually one I
can enforce in ND.

This PR implements merging of multi-parent charges. I tried for a bit to
reuse the multi-parent merging code that we use in normal ingest
mappings, but it makes assumptions about not having any backedges set
which I didn't want to work around and it was easier to just write
something fresh since this code is all going away when we delete legacy
sentences.

Ran a non-crashing ND pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-09-06_08_46_06-2823750479490581821;graphView=0?project=recidiviz-staging&authuser=0&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
and confirmed that all sentence tables produce identical results to
legacy normalization.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29517

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5f08fc21de0f35fe9593bc4cc5817f4c8f6bdba3