[admin panel] update flashing checklist to work with antd 4.24 (Recidiviz/recidiviz-data#33075)
## Description of the change

Recently we learned that the flashing checklist is broken due to an antd
upgrade from january; this PR updates the flashing checklist to work w/
the antd library updates. here's a quick video of me clicking through
it. i confirmed that a few of the action buttons worked (updated queue
status successfully) but didnt click all the way through the checklist
since OZ doesnt work w/ this version of the checklist really (raw data
import dag is enabled)

https://github.com/user-attachments/assets/4360e4c4-994e-4b16-ab8b-07fc49385260

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#32449

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f3f4f500f0451851e2dbb971e51b2751ed037608