Applies lower case to all workflows usage email addresses (Recidiviz/recidiviz-data#33143)
## Description of the change

We were seeing abnormally low numbers of staff with active usage in TN &
ID in the new usage methodology that joins usage events to primary usage
sessions. This was because of mismatched capitalization in usage event
emails, so this change makes sure all entries in
`workflows_officer_events` have lower case.

The following query shows 280 staff members who had a usage event in the
past month but did not join to product roster archive -- after checking
sandbox, the count goes to 0:
```
WITH active_users AS (
  SELECT DISTINCT
    state_code,
    email_address
  FROM
    `recidiviz-123.analyst_data.workflows_user_events_materialized`
  WHERE event in ("WORKFLOWS_USER_ACTION", "WORKFLOWS_USER_PAGE", "WORKFLOWS_USER_CLIENT_STATUS_UPDATE")
  AND DATE_TRUNC(event_date, MONTH) = "2024-08-01"
)
SELECT
  *
FROM
  active_users a
LEFT JOIN
  `recidiviz-123.mayukas_analyst_data.workflows_user_product_roster_archive_sessions_materialized` b
ON
  a.state_code = b.state_code
  and a.email_address = b.workflows_user_email_address
  and b.end_date_exclusive is null
WHERE b.workflows_user_email_address IS NULL
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: af0fef983900d1a225ba38ea76206207a8a45eab