[US_AZ] Add DPR fields to latest AZ_DOC_SC_OFFENSE view (Recidiviz/recidiviz-data#33145)
## Description of the change

There are many datetime fields in this raw data table related to the
drug transition program that we were not including in the `_latest` raw
data view initially because they were not relevant to our goals, and the
table was 200 columns wide without them. Now that we are working on
adding DPR to our TPR tool, it may be useful to have these fields
accessible in the `_latest` view. This PR add descriptions and datetime
parsers to the raw data config so these fields will be included.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f677ab741e78d49a187939b82d59924a953c4f27