[Ingest] Fix backedge validation for NormalizedStateEarlyDischarge (Recidiviz/recidiviz-data#33176)
## Description of the change

The combined ingest + normalization pipeline for US_IX crashes because
we were trying to enforce that BOTH `incarceration_sentence` and
`supervision_sentence` are nonnull backedges on the
`NormalizedStateEarlyDischarge`.

This change allows them to be null because only at most one of those
values can be set.

Pipeline run with these changes
[here](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-09-09_16_48_49-12395173803181895562;graphView=0?project=recidiviz-staging&authuser=0&organizationId=448885369991&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))).
(EDIT: this pipeline failed due to an unrelated issue in main, will
update with a new pipeline when I can).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29517

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c02b7643ad6d2fad75ff3d4cc27689e54ae7369b